requirements.txt
Django==3.2.16

# asgiref==3.7.2
# setuptools==65.5.0
# sqlparse==0.4.4

djangorestframework==3.14.0
# djangorestframework-simplejwt==5.3.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.0

# drf-extra-fields==3.7.0
Pillow==10.1.0

# reportlab==4.0.7
# django-filter==23.3

# getenv==0.2.0
# python-dotenv==1.0.0
# django-cors-headers==3.13.0
# psycopg2-binary==2.9.9

# pep8==1.7.1
# django-colorfield==0.10.1